+++ src/lib/Button/index.less
@@ -1,15 +1,39 @@
 @import '../../style/index.less';
 
+
+
 .kl-button {
   border: none;
   cursor: pointer;
-  border-radius: 4px;
+  border-radius: @button-radius;
   border: 1px solid;
   border-color: #eee;
   outline: none;
-  transition: background-color .1s ease, border-color .1s ease, box-shadow .2s ease;
+  transition: background-color .25s ease, border-color .25s ease, box-shadow .25s ease;
   box-shadow: 0px 2px 10px 0px rgba(0, 0, 0, .12);
   position: relative;
+
+  .button--children {
+    position: relative;
+    z-index: 1;
+  }
+
+  .kl-ripple {
+    left: -1px;
+    top: -1px;
+    width: calc(100% + 2px);
+    height: calc(100% + 2px);
+    border-radius: @button-radius;
+  }
+
+  &.button--border {
+    .kl-ripple {
+      left: 0;
+      top: 0;
+      width: 100%;
+      height: 100%;
+    }
+  }
 
   &:hover {
     box-shadow: 0px 3px 14px 2px rgba(0, 0, 0, .12)
@@ -19,14 +43,108 @@
     box-shadow: 0px 5px 16px 6px rgba(0, 0, 0, .12)
   }
 
-  .kl-ripple{
-    left: -1px;
-    top: -1px;
-    width: calc(100% + 2px);
-    height: calc(100% + 2px);
+  .button-size-mixin();
+
+
+  //style
+  &.button--default {
+    color: @font-color-default;
+    border-color: @color-default;
+    background-color: @color-default;
+
+    &:hover {
+      border-color: darken(@color-default, 10%);
+      background-color: darken(@color-default, 10%);
+    }
+
+
+    &.button--border {
+      .button--border-mixin(@color-default, darken(@color-default, 10%), lighten(@color-default, 8%), @color-default);
+      color: @font-color-default;
+    }
   }
 
-  // 32 40 48
+  &.button--primary {
+    color: @font-color-primary;
+    border-color: @color-primary;
+    background-color: @color-primary;
+
+    &:hover {
+      border-color: darken(@color-primary, 10%);
+      background-color: darken(@color-primary, 10%);
+    }
+
+    &.button--border {
+      .button--border-mixin(lighten(@color-primary, 10%), @color-primary, lighten(@color-primary, 40%), lighten(@color-primary, 30%));
+    }
+  }
+
+  &.button--success {
+    color: @font-color-success;
+    border-color: @color-success;
+    background-color: @color-success;
+
+    &:hover {
+      border-color: darken(@color-success, 10%);
+      background-color: darken(@color-success, 10%);
+    }
+
+    &.button--border {
+      .button--border-mixin(lighten(@color-success, 10%), @color-success, lighten(@color-success, 45%), lighten(@color-success, 25%));
+    }
+  }
+
+  &.button--info {
+    color: @font-color-info;
+    border-color: @color-info;
+    background-color: @color-info;
+
+    &:hover {
+      border-color: darken(@color-info, 10%);
+      background-color: darken(@color-info, 10%);
+    }
+
+    &.button--border {
+      .button--border-mixin(lighten(@color-info, 10%), @color-info, lighten(@color-info, 40%), lighten(@color-info, 30%));
+    }
+  }
+
+  &.button--error {
+    color: @font-color-error;
+    border-color: @color-error;
+    background-color: @color-error;
+
+    &:hover {
+      border-color: darken(@color-error, 10%);
+      background-color: darken(@color-error, 10%);
+    }
+
+    &.button--border {
+      .button--border-mixin(lighten(@color-error, 10%), @color-error, lighten(@color-error, 50%), lighten(@color-error, 35%));
+    }
+  }
+
+  &.button--warn {
+    color: @font-color-warn;
+    border-color: @color-warn;
+    background-color: @color-warn;
+
+    &:hover {
+      border-color: darken(@color-warn, 10%);
+      background-color: darken(@color-warn, 10%);
+    }
+
+    &.button--border {
+      .button--border-mixin(lighten(@color-warn, 10%), @color-warn, lighten(@color-warn, 30%), lighten(@color-warn, 20%));
+      color: @font-color-warn;
+    }
+  }
+}
+
+
+// 32 40 48
+.button-size-mixin {
+
   &.button-size--s {
     font-size: 13px;
     line-height: 20px;
@@ -44,71 +162,24 @@
     line-height: 24px;
     padding: 10px 24px;
   }
+}
 
-  //style
-  &.button--default {
-    color: @font-color-default;
-    border-color: @color-default;
-    background-color: @color-default;
 
-    &:hover {
-      border-color: darken(@color-default, 10%);
-      background-color: darken(@color-default, 10%);
-    }
+.button--style-mixin() {}
+
+// color同@border-hover
+.button--border-mixin(@border, @border-hover, @bg-hover, @ripple) {
+  background-color: transparent;
+  box-shadow: none;
+  color: @border-hover;
+  border-color: @border;
+
+  .kl-ripple--root {
+    background-color: @ripple;
   }
 
-  &.button--primary {
-    color: @font-color-primary;
-    border-color: @color-primary;
-    background-color: @color-primary;
-
-    &:hover {
-      border-color: darken(@color-primary, 10%);
-      background-color: darken(@color-primary, 10%);
-    }
-  }
-
-  &.button--success {
-    color: @font-color-success;
-    border-color: @color-success;
-    background-color: @color-success;
-
-    &:hover {
-      border-color: darken(@color-success, 10%);
-      background-color: darken(@color-success, 10%);
-    }
-  }
-
-  &.button--info {
-    color: @font-color-info;
-    border-color: @color-info;
-    background-color: @color-info;
-
-    &:hover {
-      border-color: darken(@color-info, 10%);
-      background-color: darken(@color-info, 10%);
-    }
-  }
-
-  &.button--error {
-    color: @font-color-error;
-    border-color: @color-error;
-    background-color: @color-error;
-
-    &:hover {
-      border-color: darken(@color-error, 10%);
-      background-color: darken(@color-error, 10%);
-    }
-  }
-
-  &.button--warn {
-    color: @font-color-warn;
-    border-color: @color-warn;
-    background-color: @color-warn;
-
-    &:hover {
-      border-color: darken(@color-warn, 10%);
-      background-color: darken(@color-warn, 10%);
-    }
+  &:hover {
+    border-color: @border-hover;
+    background-color: @bg-hover;
   }
 }
