+++ src/lib/Ripple/index.less
@@ -6,12 +6,17 @@
   .ab-cover();
   overflow: hidden;
 
+  &--cover{
+    .ab-cover();
+    z-index: 100;
+  }
+
   &--root {
     position: relative;
     transition: transform .25s ease-in, opacity .15s linear;
     border-radius: 50%;
     transform: translate3d(-50%, -50%, 0) scale(0.65, 0.65);
-    background-color: rgba(255, 255, 255, .4);
+    background-color: rgba(255, 255, 255, .5);
     // background-color: rgba(0, 0, 0, .5);
     display: block;
     opacity: 0;
