+++ src/style/variable.less
@@ -23,6 +23,10 @@
 @font-color-default: #333;
 
 
+//button 
+@button-radius:4px;
+
+
 //position
 
 .ab-cover {
